+++ src/api/configure.ts
@@ -1,5 +1,4 @@
 import { globalState, isolateGlobalState } from "../core/globalstate"
-import { reserveArrayBuffer } from "../types/observablearray"
 import { setReactionScheduler } from "../core/reaction"
 
 export function configure(options: {
@@ -7,7 +6,6 @@ export function configure(options: {
     computedRequiresReaction?: boolean
     isolateGlobalState?: boolean
     disableErrorBoundaries?: boolean
-    arrayBuffer?: number
     reactionScheduler?: (f: () => void) => void
 }): void {
     if (options.enforceActions !== undefined) {
@@ -26,9 +24,6 @@ export function configure(options: {
                 "WARNING: Debug feature only. MobX will NOT recover from errors if this is on."
             )
         globalState.disableErrorBoundaries = !!options.disableErrorBoundaries
-    }
-    if (typeof options.arrayBuffer === "number") {
-        reserveArrayBuffer(options.arrayBuffer)
     }
     if (options.reactionScheduler) {
         setReactionScheduler(options.reactionScheduler)